src/plugin-installer-job/requirements.txt
google-cloud-storage
google-cloud-secret-manager
PyYAML
